+++ components/layout.tsx
@@ -7,7 +7,6 @@ import AuthModal from "./auth/AuthModal"
 import { DesktopNav } from "./DesktopNav"
 import PageFooter from "./Footer/Footer"
 import { MobileNav } from "./MobileNav"
-import { MobileNavV2 } from "./MobileNavV2"
 
 export const PageContainer: FC<React.PropsWithChildren<unknown>> = ({
   children
@@ -46,7 +45,7 @@ export const Layout: React.FC<React.PropsWithChildren<LayoutProps>> = ({
             <link rel="icon" href="/favicon.ico" />
           </Head>
           <PageContainer>
-            {isMobile ? <MobileNavV2 /> : <DesktopNav />}
+            {isMobile ? <MobileNav /> : <DesktopNav />}
             <AuthModal />
             <div className={`col`}>{children}</div>
             <PageFooter
